fixed bug where icnn min max bounds were not assigned

+++ script/settings.py
@@ -1,4 +1,4 @@
 import torch
 
 device = "cpu"
-data_type = torch.float64
+data_type = torch.float32
